requirements.txt
Flask==1.1.2
flask-cors==3.0.9
flask_restful==0.3.8
dataclasses==0.7
oslash==0.5.1
pymongo==3.11.0
python-dotenv==0.14.0
dnspython==2.0.0
